Refactored large swaths of the code to abstract each component layer
UI is now completely separate from the Business Logic which is separate from the Data Access layer. UI does not communicate with the Data Access layer (db) except through the BL layer working as an intermediary.

Project is now compliant with proper software architecture model.

THIS BRANCH IS IN BETA. MANY THINGS ARE LIKELY TO BUG WITH IT STILL. YOU HAVE BEEN WARNED